requirements.txt
absl==0.0
absl_py==1.3.0
imageio==2.22.4
imgaug==0.4.0
keras==2.10.0
matplotlib==3.6.3
numpy==1.21.5
opencv_python_headless==4.6.0.66
pandas==1.5.3
pascal_voc_writer==0.1.4
Pillow==9.4.0
plotly==5.11.0
requests==2.25.1
scikit_learn==1.2.1
six==1.16.0
tensorflow==2.11.0
tensorflow_gpu==2.10.0
torch==1.13.1
tqdm==4.64.1
fathomnet==0.5.2
